Changed the names of filters to follow the detector nomenclature
