+++ config/filterchain/simple_pipe45.yaml
@@ -88,7 +88,7 @@ filters:
         value: 0
       - name: Max_value
         value: 100
-  - name: PipeDetectorAngle
+  - name: PipeAngleDetector
     parameters:
       - name: Enable
         value: 1
