+++ src/proc_image_processing/filters/vampire_torpedoes_close_detector.h
@@ -1,10 +1,10 @@
-/// \author csauvain
-/// \date 20/07/19
-/// TODO: Refactor code repetition
+//
+// Created by csauvain on 20/07/19.
+// Il y a des repetitions de code a ameliorer
+//
 
-
-#ifndef PROC_IMAGE_PROCESSING_VAMPIRE_TORPEDOES_DETECTOR_CLOSE_H
-#define PROC_IMAGE_PROCESSING_VAMPIRE_TORPEDOES_DETECTOR_CLOSE_H
+#ifndef PROC_IMAGE_PROCESSING_VAMPIRE_TORPEDOES_CLOSE_DETECTOR_H
+#define PROC_IMAGE_PROCESSING_VAMPIRE_TORPEDOES_CLOSE_DETECTOR_H
 
 
 #include <proc_image_processing/filters/filter.h>
@@ -26,22 +26,21 @@ namespace proc_image_processing {
         //============================================================================
         // P U B L I C   C / D T O R S
 
-        explicit VampireTorpedoesDetectorClose(const GlobalParamHandler& globalParams)
-            : Filter(globalParams),
-            enable_("Enable", false, &parameters_),
-            debug_contour_("Debug_contour", false, &parameters_),
-            look_for_ellipse_("Look_for_Ellipse", false, &parameters_),
-            look_for_heart_("Look_for_Heart", false, &parameters_),
-            min_area_("Min_area", 5000, 1, 50000, &parameters_),
-            max_area_("Max_area", 100000, 1, 1000000, &parameters_) {
-            SetName("VampireTorpedoesDetectorClose");
-        }
+        explicit VampireTorpedoesDetectorClose(const GlobalParamHandler &globalParams)
+                : Filter(globalParams),
+                enable_("Enable", false, &parameters_),
+                debug_contour_("Debug_contour", false, &parameters_),
+                look_for_ellipse_("Look_for_Ellipse", false, &parameters_),
+                look_for_heart_("Look_for_Heart", false, &parameters_),
+                min_area_("Min_area", 5000, 1, 50000, &parameters_),
+                max_area_("Max_area", 100000, 1, 1000000, &parameters_) {
+            SetName("VampireTorpedoesDetectorClose");}
 
         virtual ~VampireTorpedoesDetectorClose() {}
 
         //============================================================================
         // P U B L I C   M E T H O D S
-        virtual void Execute(cv::Mat& image) {
+        virtual void Execute(cv::Mat &image){
             if (enable_()) {
                 std::string objectif;
                 image.copyTo(output_image_);
@@ -84,7 +83,7 @@ namespace proc_image_processing {
                     }
 
                     cv::Mat pointfs;
-                    cv::Mat(contours[i]).convertTo(pointfs, CV_32F);
+                    cv::Mat(contours[i]).convertTo(pointfs,CV_32F);
                     cv::RotatedRect box = cv::fitEllipse(pointfs);
 
                     float circleIndex;
@@ -148,7 +147,7 @@ namespace proc_image_processing {
                     target.SetTarget(objectif, center.x, center.y, object->GetWidth(), object->GetHeight(), object->GetRotatedRect().angle, image.rows, image.cols);
                     NotifyTarget(target);
                     if (debug_contour_()) {
-                        cv::circle(output_image_, objVec[0]->GetCenter(), 3, CV_RGB(0, 255, 0), 3);
+                        cv::circle(output_image_, objVec[0]->GetCenter(), 3, CV_RGB(0,255,0),3);
                     }
                 }
 
@@ -157,7 +156,6 @@ namespace proc_image_processing {
                 }
             }
         }
-
 
     private:
         //============================================================================
@@ -172,4 +170,4 @@ namespace proc_image_processing {
 
 }  // namespace proc_image_processing
 
-#endif  // PROC_IMAGE_PROCESSING_VAMPIRE_TORPEDOES_DETECTOR_CLOSE_H
+#endif  // PROC_IMAGE_PROCESSING_VAMPIRE_TORPEDOES_CLOSE_DETECTOR_H
